add Pentogram::draw

+++ headers/pentagram_builder.h
@@ -12,7 +12,7 @@ _DRAWING_BEGIN
 REG_BUILDER_BEGIN(Pentagram)
 
     BUILDER(Pentagram)& setup_center(const Point &_c);
-    BUILDER(Pentagram)& setup_radius(std::size_t _r);
+    BUILDER(Pentagram)& setup_radius(std::uint64_t _r);
     BUILDER(Pentagram)& setup_thickness(std::size_t _t);
     BUILDER(Pentagram)& setup_color(const png::Color &_c);
